% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alb_capitals}
\alias{alb_capitals}
\title{Geographic coordinates of the Albanian capitals}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12 rows and 9 columns.
}
\source{
\url{https://data.humdata.org/dataset/albania-administrative-level-0-3-boundaries}
}
\usage{
alb_capitals
}
\description{
A data frame with latitude, longitude and name of the capital cities of the
administrative regions of Albania.
}
\keyword{datasets}
