% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Coerciones implicitas}
\source{
http://bit.ly/Instructivo_package_IntRo
}
\usage{
ci(type = c("logical", "integer", "double", "raw", "complex", "character"))
}
\arguments{
\item{type}{vector de caracteres mayor o igual a dos que indica el tipo de datos
se desea testear.}
}
\description{
Esta funcion devuelve el resultado del tipo de datos resultante
    de combinar en un vector atomico datos de distinto tipo. Los tipos posibles
    de datos que puede tener un vector atomico son:

    \itemize{
    \item{\code{logical}}
    \item{\code{integer}}
    \item{\code{double}}
    \item{\code{raw}}
    \item{\code{complex}}
    \item{\code{character}}
    }

    Cuando se combinan datos de distinto tipo R resuelve coercionar el vector
    hacia uno de los tipos que se estan combianndo. Este resultado no es aleatorio
    sino que es resultado de una regla clara en el lenguaje sobre 'jerarquia' de
    tipos de datos. La funcion con la que se verifica el tipo de dato de un
    vector atomico es \code{typeof()}
}
\examples{
ci()
ci(c("character", "numeric"))
}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
