% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_intRo.R
\name{deploy_intRo}
\alias{deploy_intRo}
\title{Deploy an instance of intRo to ShinyApps.io}
\usage{
deploy_intRo(path = getwd(), enabled_modules = NULL, theme = NULL,
  google_analytics = NULL, ...)
}
\arguments{
\item{path}{The directory containing the intRo folder}

\item{enabled_modules}{The modules to enable}

\item{theme}{The shinythemes theme to use}

\item{google_analytics}{The google analytics tracking ID to use}

\item{...}{Additional arguments passed to the shinyapps deployApp function}
}
\description{
Deploy an instance of intRo to ShinyApps.io
}
\examples{
\dontrun{
    deploy_intRo()
                 theme = "cerulean")
}
}

