% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_ratios}
\alias{df_ratios}
\title{Student/teacher ratio by country}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 180 rows and 20 columns.
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
df_ratios
}
\description{
A data frame containing the student-to-teacher ratio in the world's countries.
}
\keyword{datasets}
