% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_intRo.R
\name{download_intRo}
\alias{download_intRo}
\title{Download an instance of intRo to your machine}
\usage{
download_intRo(path = getwd())
}
\arguments{
\item{path}{The directory to download intRo to}
}
\description{
Download an instance of intRo to your machine
}
\examples{
\dontrun{
    download_intRo()
}
}

