% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{ejercicios}
\alias{ejercicios}
\title{Pauta de ejercicios del curso IntRo}
\usage{
ejercicios(modulo = 1)
}
\arguments{
\item{modulo}{valor numerico que indica el repartido de ejercicios que se desea consultar.
\describe{
         \item{1}{Vectores}
         \item{2}{Matrices}
         \item{3}{Data Frame}
         \item{4}{Listas}
        }}
}
\value{
Abre un archivo .pdf en el lector de pdf que se tenga instalado.
}
\description{
Permite acceder a los repartidos de ejercicios de cada modulo del curso
}
\details{
Si se pone por valor del argumento \code{modulo} un numero fuera del rango se va a
    abrir la carpeta en la que estan los ejercicios y desde ahi poder seleccionar la que se desea.
}
\examples{
# ejercicios()
# ejercicios(modulo = 1)
}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
