% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{endangered}
\alias{endangered}
\title{Endangerment level of the world's languages}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7845 rows and 6 columns.
}
\source{
\url{https://github.com/glottolog/glottolog-cldf}
}
\usage{
endangered
}
\description{
A data frame containing the a list of languages and their endangerment status.
}
\keyword{datasets}
