% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escala.R
\name{escala}
\alias{escala}
\title{Escala de notas.}
\usage{
escala(puntaje)
}
\arguments{
\item{puntaje}{Numero entre 0 y 100.}
}
\value{
Numero entero entre 0 y 12.
}
\description{
Esta funcion contiene la escala de notas. Convierte los puntajes
    de los ejercicios a la escala de notas de 0 a 12.
}
\examples{

escala(82)
escala(88.5)

}
