% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{europe}
\alias{europe}
\title{Europe}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 51 rows and 95 columns.
}
\source{
\url{https://www.naturalearthdata.com}
}
\usage{
europe
}
\description{
Political boundaries of European countries.
}
\keyword{datasets}
