% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harry_potter}
\alias{harry_potter}
\title{Astrological sign and Hogwarts houses}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 447 rows and 3 columns.
}
\source{
\url{https://imgur.com/gallery/wbuRF}
}
\usage{
harry_potter
}
\description{
This table contains responses from 447 people who reported they assigned
Hogwarts house and their birth sign.
}
\keyword{datasets}
