% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{intro_tutorial}
\alias{intro_tutorial}
\title{Run tutorials}
\usage{
intro_tutorial()
}
\value{
Nothing. Used for its side effects.
}
\description{
The function asks you which tutorial to open and then runs it.
}
\examples{
# run_tutorial()
}
