% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_use}
\alias{land_use}
\title{Use of land in the world through the years}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 159 rows and 3 columns.
}
\source{
\url{https://ourworldindata.org/land-use}
}
\usage{
land_use
}
\description{
A table with information on the type of land use and the land area over time.
}
\keyword{datasets}
