% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_use_untidy}
\alias{land_use_untidy}
\title{Use of land in the world through the years (untidy)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 53 rows and 4 columns.
}
\source{
\url{https://ourworldindata.org/land-use}
}
\usage{
land_use_untidy
}
\description{
A table with information on the type of land use and the land area over time.
}
\keyword{datasets}
