% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{manual}
\alias{manual}
\title{Manual de envio de ejercicio y paquete}
\usage{
manual(manual = "envio")
}
\arguments{
\item{manual}{valor de tipo \code{character} que por defecto es \code{enviar} y admite otro valor que es
el de \code{paquete}.}
}
\value{
Abre un archivo .pdf en el lector de pdf que se tenga instalado.
}
\description{
Permite acceder al manual del procedimiento de envio por mail desde R de los ejercicios
    una vez finalizados. Tambien permite acceder al manual del paquete y a la documentacion de todas las
    funciones del este.
}
\examples{
# manual()
# manual(manual = "paquete")
}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
