% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{messy_fruit}
\alias{messy_fruit}
\title{An untidy fruity mess}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 4 columns.
}
\source{
Stefano Coretta
}
\usage{
messy_fruit
}
\description{
An inventory of sold and bought fruit, but untidy.
}
\keyword{datasets}
