% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{person_united}
\alias{person_united}
\title{Top Personality Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1834 rows and 2 columns.
}
\source{
Arslan Ali, \url{https://www.kaggle.com/arslanali4343/top-personality-dataset}
}
\usage{
person_united
}
\description{
(MBTI) Myers-Briggs Personality Type Dataset. The scores are in a single column.
}
\keyword{datasets}
