% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{presentaciones}
\alias{presentaciones}
\title{Presentaciones del curso}
\usage{
presentaciones(modulo = 0)
}
\arguments{
\item{modulo}{valor numerico que indica el la presentacion de que modulo se desea abrir.
\describe{
         \item{0}{Preliminares}
         \item{1}{¿Que es R?}
         \item{2}{Fundamentos de programacion en R}
         \item{3}{Vectores}
         \item{4}{Matrices}
         \item{5}{Data Frame}
         \item{6}{Listas}
         \item{7}{Graficos}
         \item{8}{Sentencias de control y Funciones}
        }}
}
\value{
Abre un archivo .pdf en el lector de pdf que se tenga instalado.
}
\description{
Permite acceder a las presentaciones (slides) del docente en formato .pdf
}
\details{
Si se pone por valor del argumento \code{modulo} un numero fuera del rango se va a
    abrir la carpeta en la que estan las presentaciones y desde ahi poder seleccionar la que se desea.
}
\examples{
# presentaciones()
# presentaciones(modulo = 1)
}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
