% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pyramids}
\alias{pyramids}
\title{The pyramids of Egypt}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 62 rows and 16 columns.
}
\source{
Daria Chemkaeva, \url{https://www.kaggle.com/lsind18/egyptianpyramids}
}
\usage{
pyramids
}
\description{
Information on the pyramids of Egypt.
}
\keyword{datasets}
