% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunction.R
\name{rfunction}
\alias{rfunction}
\title{Funcion de busqueda aleatoria de funciones de uno o todos los paquetes
    instalados.}
\source{
http://bit.ly/Instructivo_package_IntRo
}
\usage{
rfunction(package = "base", ALL = FALSE)
}
\arguments{
\item{package}{por defecto en la funcion esta cargado el paquete \code{base}.
Pero se puede seleccionar cualquiera de los paquetes que fueron instalados
en su computadora.}

\item{ALL}{por defecto es \code{FALSE}. Si es \code{TRUE} significa que la busqueda
aleatoria se va a buscar entre todos los paquetes que estan cargados en la
sesion de trabajo actual.}
}
\value{
Esta funcion va a imprimir en pantalla la funcion sorteada, el pquete al
    que pertenece y la antidad de funciones que ese paquqte tiene. Adicionalmente
    se va a abrir un device de la ayuda de la funcion sorteada.
}
\description{
Esta funcion permite obtener una funcion aleatoria de alguno de los
    paquete instalados  con fines de aprendizaje. Es decir,
    la idea es seleccionar al azar una funcion y la funcion que salga sorteada
    es con la que se debe experimentar para aprender sobre ella.
}
\examples{

rfunction()
rfunction(ALL = TRUE)
rfunction(package = "stats")


}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
