% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_intRo.R
\name{run_intRo}
\alias{run_intRo}
\title{Run a local instance of intRo}
\usage{
run_intRo(path = getwd(), enabled_modules = NULL, theme = NULL)
}
\arguments{
\item{path}{The directory containing the intRo folder}

\item{enabled_modules}{The modules to enable}

\item{theme}{The shinythemes theme to use}
}
\description{
Run a local instance of intRo
}
\examples{
\dontrun{
    run_intRo()
    run_intRo(enabled_modules = c("data/transform", "summaries/graphical"),
              theme = "cerulean")
}
}

