% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_tutorial}
\alias{run_tutorial}
\title{Run tutorials}
\usage{
run_tutorial()
}
\value{
Nothing. Used for its side effects.
}
\description{
The function asks you which tutorial to open and then runs it.
}
\examples{
# run_tutorial()
}
