% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_options.R
\name{set_options}
\alias{set_options}
\title{Set the options for the given intRo instance}
\usage{
set_options(path, enabled_modules, theme)
}
\arguments{
\item{path}{The directory containing the intRo folder}

\item{enabled_modules}{The modules to enable}

\item{theme}{The shinythemes theme to use}
}
\description{
Set the options for the given intRo instance
}
\examples{
\dontrun{
    deploy_intRo()
}
}

