% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slavery_dis}
\alias{slavery_dis}
\title{Number of enslaved people during colonisation.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 63 rows and 3 columns.
}
\source{
\url{https://slavevoyages.org/voyage/about#methodology/introduction/0/en/}
}
\usage{
slavery_dis
}
\description{
This data frame holds information on the number of enslaved people who were
brought from Africa to the New World.
}
\keyword{datasets}
