% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slavery_tot}
\alias{slavery_tot}
\title{Number of enslaved people during colonisation.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\source{
\url{https://slavevoyages.org/voyage/about#methodology/introduction/0/en/}
}
\usage{
slavery_tot
}
\description{
This data frame holds information on the number of enslaved people who were
brought from Africa to the New World.
}
\keyword{datasets}
