% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{social}
\alias{social}
\title{Number of users of social media through the years}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 142 rows and 3 columns.
}
\source{
\url{https://ourworldindata.org/rise-of-social-media}
}
\usage{
social
}
\description{
A data table with number of users for major social media through the years.
}
\keyword{datasets}
