% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yoda_corpus}
\alias{yoda_corpus}
\title{Yoda Speech Corpus}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 371 rows and 7 columns.
}
\source{
Stefano Coretta, \url{https://www.kaggle.com/stefanocoretta/yoda-speech-corpus}
}
\usage{
yoda_corpus
}
\description{
Corpus of dialogues from the Star Wars movies with Yoda as an interlocutor.
}
\keyword{datasets}
