% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_eval_vector.R
\name{auto_eval_vector}
\alias{auto_eval_vector}
\title{Autoevaluacion de 'Ejercicios I' (Modulo Vectores).}
\source{
http://bit.ly/Instructivo_package_IntRo
}
\usage{
auto_eval_vector(
  nombre = NULL,
  apellido = NULL,
  mail = NULL,
  intentos = 1,
  enviar = FALSE,
  summary = TRUE
)
}
\arguments{
\item{nombre}{Una cadena de caracteres que indique el nombre del estudiante.}

\item{apellido}{Una cadena de caracteres que indique el apellido del estudiante.}

\item{mail}{Una cadena de caracteres que indique el mail del estudiante.}

\item{intentos}{Refiere a la cantidad de veces que se realizo la autoevaluacion
hasta el momento en el que se resuelve que los ejercicios estan prontos para
ser enviados para obtener la nota final. Por defeto es 1. La cantidad de
intentos debe ser un numero entero. Si el numero no es entero la funcion
va a forzar el numero a uno de tipo  \code{integer} redondeando hacia
arriba.}

\item{enviar}{Valor logico que por defeto es \code{FALSE}. En el momento en el que el
estudiante considera que los ejercicios estan completos debe enviar los
resultados para obtener la nota final de los ejercicios. La nota final le va
a llegar al mail que cargo en el argumento \code{mail} de esta funcion.}

\item{summary}{Valor logico que por defetco es \code{TRUE}. La utilidad de este
argumento es contar con un resumen de la situacion de todos los ejercicios
correspondientes al modulo de vectores.}
}
\value{
\itemize{

    \item{Si el argumento \code{summary} es \code{TRUE}}{ la funcion va a devolver
    un \code{print} del estado de los ejercicios. El resultado de los ejercicioes
    pueden estar en tres categorias: respuesta 'Correcta', ' Incorrecta' o '
    Incompleta'.}
    \item{Si el argumento \code{summary} es \code{FALSE}}{ la funcion no va a imprimir
    en la consola ninun resulatdos. Solo se van a ver los Errores y las Advertencias
    si son necesarias.}
    }
}
\description{
Es una funcion que permite autoevaluar los ejercicios del modulo
    sobre vectores del curso Introduccion al software estadistico R.
}
\examples{

vec1 <- seq(2, 20, 2)

auto_eval_vector(nombre    = "Nicolas",
                 apellido  = "Schmidt",
                 mail      = NULL,
                 enviar    = FALSE,
                 intentos  = 3,
                 summary   = TRUE)

## --------------------
## ENVIAR EL EJERCICIO
## --------------------

# auto_eval_vector(nombre   = "Nicolas",
#                  apellido = "Schmidt",
#                  mail     = "tu_mail.gmail.com",  ## VALOR NECESARIO!
#                  enviar   = TRUE,                 ## VALOR NECESARIO!
#                  intentos = 3,
#                  summary  = FALSE)

}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
