% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deploy_intRo.R
\name{deploy_intRo}
\alias{deploy_intRo}
\title{Deploy an instance of intRo to ShinyApps.io}
\usage{
deploy_intRo(path = getwd(), ...)
}
\arguments{
\item{path}{The directory containing the intRo folder}

\item{...}{Additional arguments passed to the shinyapps deployApp function}
}
\description{
Deploy an instance of intRo to ShinyApps.io
}
\examples{
\dontrun{
    deploy_intRo()
}
}

