% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunction.R
\name{rfunction}
\alias{rfunction}
\title{Busqueda aleatoria de funciones de uno o todos los paquetes
    instalados.}
\source{
http://bit.ly/Instructivo_package_IntRo
}
\usage{
rfunction(package = "base", ALL = FALSE, library = TRUE)
}
\arguments{
\item{package}{por defecto en la funcion esta cargado el paquete \code{base}.
Pero se puede seleccionar cualquiera de los paquetes que fueron instalados
en su computadora.}

\item{ALL}{por defecto es \code{FALSE}. Si es \code{TRUE} significa que la busqueda
aleatoria se va a buscar entre todos los paquetes que estan cargados en la
sesion de trabajo actual. Si adicionalmente se selecciono un paquete en
\code{package} este sera incorporado al muestreo.}

\item{library}{por defecto es \code{TRUE}. Esto significa que el paquete seleccionado
queda cargado en la rurta de busqueda de nombres de R. Caso contrario el paquete no
va a quedar cargado.}
}
\value{
Esta funcion va a imprimir en pantalla la funcion sorteada, el paquete al
    que pertenece y la cantidad de funciones que ese paquqte tiene. Adicionalmente
    se va imprimir una descripcion del paquete y se va a abrir un device de la
    ayuda de la funcion sorteada.
}
\description{
Esta funcion permite obtener una funcion aleatoria de alguno de los
    paquete instalados  con fines de aprendizaje. Es decir,
    la idea es seleccionar al azar una funcion y la funcion que salga sorteada
    es con la que se debe experimentar para aprender sobre ella.
}
\details{
Si se selecciona un paquete que no esta cargado en la sesion
    de trabajo no se afecta la ruta de busqueda de nombres de R. El
    paquete seleccionado no va a quedar cargado en la ruta actual unicamente
    si selecciona  \code{library = FALSE}.Si se selecciona un paquete que
    no esta instalado la funcion le va a preguntar si desea instalarlo mediante
    una ventana emergente. Si no desea instalar o la instalacion es fallida por
    algun motivo la funcion va a debolver un muestreo del paquete 'base'.
}
\examples{

rfunction()
rfunction(ALL = TRUE)
rfunction(package = "stats")


}
\author{
Nicolas Schmidt \email{nschmidt@cienciassociales.edu.uy}
}
